WIP: Beginning to define file/csv store
